requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==3.2.14
django-heroku==0.3.1
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
psycopg2==2.9.3
pymongo==3.12.3
pytz==2022.1
sqlparse==0.2.4
tzdata==2022.1
whitenoise==6.2.0



